requirements.txt
psycopg2-binary==2.9.11
requests==2.32.5
pandas==2.3.3
streamlit==1.41.1
plotly==6.4.0
